socialauthentication

+++ controllers/api/v1/posts_api.js
@@ -1,21 +1,21 @@
-const Post=require('../../../models/post');
-const Comment=require('../../../models/comment');
+const Post = require('../../../models/post');
+const Comment = require('../../../models/comment');
 
-module.exports.index=async function(req,res){
+module.exports.index = async function (req, res) {
 
     let posts = await Post.find({})
-         .sort('-createdAt')
-         .populate('user')
-         .populate({
+        .sort('-createdAt')
+        .populate('user')
+        .populate({
             path: 'comments',
             populate: {
-               path: 'user'
+                path: 'user'
             }
-         });
+        });
 
-    return res.json(200,{
-        message:"List of posts",
-        posts:posts
+    return res.json(200, {
+        message: "List of posts",
+        posts: posts
     })
 }
 
@@ -24,25 +24,26 @@ module.exports.destroy = async function (req, res,) {
     try {
         let post = await Post.findById(req.params.id);
         //.id means converting the object id into string
-       // if (post.user == req.user.id) {
+        if (post.user == req.user.id) {
             post.remove();
 
             await Comment.deleteMany({ post: req.params.id });
 
-           // req.flash('success','Post and associated comments deleted!');
+            // req.flash('success','Post and associated comments deleted!');
 
-            return res.json(200,{
-                message:"Posts and associated comments deleted successfully"
+            return res.json(200, {
+                message: "Posts and associated comments deleted successfully"
             });
-        
-        // else {
-        //     req.flash('error','You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }
+        else {
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            })
+        }
     } catch (err) {
         //req.flash('error',err);
-        return res.json(500,{
-            message:"Internal Server Error"
+        return res.json(500, {
+            message: "Internal Server Error"
         });
     }
 }
